✨ feat(tmac): enhance TMAC operator and related scripts
- add environment variable for num_threads in op_gen.sh
- update architecture detection and compile command in op_gen.sh
- include necessary headers and remove unused code in op_tmac-impl.hpp
- modify preprocess and qgemm_lut functions to support new parameters and return types
- update Torch library definitions in op_tmac_aten.cpp
- adjust kernel configurations in model_utils.py
- update data types and vectorization in qgemm.py
- update test_e2e.py with new file paths and additional test data
- load specific data in test_tmac.py and add debug breakpoints
- update weight quantization and test cases in test_op_tmac.py

🔧 chore(tmac): clean up and optimize code

- remove redundant blank lines and comments in compile_mod.py
- remove unused print_tensor_data function in op_tmac-impl.hpp
- update data types and tensor operations in op_tmac-impl.hpp
- update test cases and data loading in test_tmac.py and test_op_tmac.py
- add debug breakpoints and print statements for better traceability

✅ test(tmac): update and enhance test cases

- update test_e2e.py with new file paths and additional test data
- load specific data in test_tmac.py and add debug breakpoints
- update weight quantization and test cases in test_op_tmac.py
- ensure consistency and correctness of the test cases

🌐 i18n(tmac): ensure English language consistency

- update comments and messages to be in English in op_tmac-impl.hpp
- ensure all strings and comments are in English across the changes

+++ torchao/experimental/ops/tmac/t_mac/ops/qgemm.py
@@ -199,7 +199,7 @@ class QGeMMLUTBitsCodegen(OpCodegen):
                         te.indexdiv(m, self.simd_n_out) * self.simd_n_out * self.bits
                             + te.indexmod(m, self.simd_n_out)
                             + b * self.simd_n_out
-                    ].astype("float32") * alphas[b]
+                    ].astype("float16") * alphas[b]
                     for b in range(self.bits)
                 ]),
             ).astype(self.out_dtype),
@@ -259,6 +259,8 @@ class QGeMMLUTBitsCodegen(OpCodegen):
         # 32 for vectorization should be enough
         moC, miC = sch[CC].split(mC, factor=32)
         mio, mii = sch[C].split(mi, factor=32)
+
+        #! If you generate C Code, _vectorization should be False.
         if self._vectorization:
             sch[CC].vectorize(miC)
             sch[C].vectorize(mii)
